Pass around sentiment data using the same namedtuple, making it easier to later reconfigure the data file format
