+++ scripts/sentiment/process_scare.py
@@ -22,7 +22,49 @@ import sys
 
 import stanza
 
+from scripts.sentiment.process_utils import Fragment
 import scripts.sentiment.process_utils as process_utils
+
+def get_scare_snippets(nlp, csv_dir_path, text_id_map, filename_pattern="*.csv"):
+    """
+    Read snippets from the given CSV directory
+    """
+    num_short_items = 0
+
+    snippets = []
+    csv_files = glob.glob(os.path.join(csv_dir_path, filename_pattern))
+    for csv_filename in csv_files:
+        with open(csv_filename, newline='') as fin:
+            cin = csv.reader(fin, delimiter='\t', quotechar='"')
+            lines = list(cin)
+
+            for line in lines:
+                ann_id, begin, end, sentiment = [line[i] for i in [1, 2, 3, 6]]
+                begin = int(begin)
+                end = int(end)
+                if sentiment.lower() == 'unknown':
+                    continue
+                elif sentiment.lower() == 'positive':
+                    sentiment = 2
+                elif sentiment.lower() == 'neutral':
+                    sentiment = 1
+                elif sentiment.lower() == 'negative':
+                    sentiment = 0
+                else:
+                    raise ValueError("Tell John he screwed up and this is why he can't have Mox Opal: {}".format(sentiment))
+                if ann_id not in text_id_map:
+                    print("Found snippet which can't be found: {}-{}".format(csv_filename, ann_id))
+                    continue
+                snippet = text_id_map[ann_id][begin:end]
+                doc = nlp(snippet)
+                text = " ".join(" ".join(token.text for token in sentence.tokens) for sentence in doc.sentences)
+                num_tokens = sum(len(sentence.tokens) for sentence in doc.sentences)
+                if num_tokens < 4:
+                    num_short_items = num_short_items + 1
+                snippets.append(Fragment(sentiment, text))
+    print("Number of short items: {}".format(num_short_items))
+    return snippets
+
 
 def main(in_directory, out_directory, short_name):
     os.makedirs(out_directory, exist_ok=True)
@@ -47,7 +89,7 @@ def main(in_directory, out_directory, short_name):
 
     print("Found %d total sentiment ratings" % len(text_id_map))
     nlp = stanza.Pipeline('de', processors='tokenize')
-    snippets = process_utils.get_scare_snippets(nlp, os.path.join(in_directory, "scare_v1.0.0", "annotations"), text_id_map)
+    snippets = get_scare_snippets(nlp, os.path.join(in_directory, "scare_v1.0.0", "annotations"), text_id_map)
 
     print(len(snippets))
     process_utils.write_list(os.path.join(out_directory, "%s.train.txt" % short_name), snippets)
